Replace *ngIf in BookListComponent with ngSwitch + enum
Added an enum for the view type (either List or Grid). The template
uses the ngSwitch directive to match on the enum value and switch the
correct block in and out of the DOM.

Also added a Bootstrap panel component to encapsulate the button group
and the list of books.