Emit event when book is selected from the grid
Changed the link in the book grid item to a button.  Clicking the button fires an event.
The parent grid view component listens for this event and the handler
loads the book details route for the selected book.

This serves as an example of how data can be passed up from a child component
to the parent component. This also shows how to do programmatic routing
in the code instead of placing a routerLink directive in the template.