Add bootstrap's column resets in grid view
Bootstrap's responsive column resets ensure that each new line in the
grid clears the one above it. Moved the explanation of the grid system
into the grid view component.

The grid view component also now contains an example of an alternative
syntax for ngFor that is useful in more complex situations.